Fixed tab & sidebar reload
We're properly cleaning up the page when extension is closed, by
broadcasting a message to the content scripts juste before unloading
them. Upon reception, applied helpers are reverted and the minimap is
unmounted.

We're also reapplying previous state when the page is reloaded, by
sending the app state to the content scripts each time the page finishes
loading.