Using animation frames to update the minimap
Redrawing the minimap is not costly at all, so we can get rid of
debouncing and throttling and use animation frames to rerender as soon
as possible, smoothing out the end result.
Also, the minimap is now shown whenever a test is active, whereas
helpers affect the page or not.