Improved display of fixed elements in the minimap
Element positionned relatively to the window are now displayed
accordingly in the minimap, thus avoiding weird jumps of the pins when
scrolling the page.
Their size on the map is calculated in relation to the viewport, and
they are shown in a different color so they can be distinguished from
the regular pins.