Redesigned navigation
The help page is moved into a dedicated page, allowing for a more
confortable layout. This removes the need for a markdown build.
Reference actions are moved into the header, as the reference page is
now the only page displayed in the panel. This removes the need for a
router.