Fixed extension loading on Firefox
On Firefox, the side panel stays open when switching tabs. When the
extension hasn't been loaded on a given tab yet, a button let the user
load it.

Given the current permissions (`activeTab` only), the browsers only
provides a tab's URL when requested after some select user interactions.
(i.e. a click on the browser action, see
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/permissions#activetab_permission).

Clicking the button in the panel would not register as a user
interaction, thus preventing the extension from getting a tab's URL.
This would further prevent the extension from loading altogether.

Instead of requesting privileged access to the URL through permissions
(`tabs` or host permissions), which would trigger alerts on installation
and a more thorough reviewing process, we're removing the button and
forcing the user to load the extension via the browser action.